LibAAH_RTP: Fix a deserialization bug.  DO NOT MERGE
Fix for bug 6446715

Fix a deserialization bug which was causing an assertion failure in
stagefright and bringing down the entire media server.  Basically, you
will hit this any time you play a track longer than ~2147 seconds.
(Technically, the crash will happen any time the play pointer at a
 position which is ~2147 * (1 * 2N) where N >= 0)

+++ media/libaah_rtp/aah_rx_player_substream.cpp
@@ -200,7 +200,8 @@ void AAH_RXPlayer::Substream::processPayloadStart(uint8_t* buf,
     if (ts_valid) {
         uint32_t ts_upper = U32_AT(buf + parse_offset);
         parse_offset += 4;
-        ts = (static_cast<int64_t>(ts_upper) << 32) | ts_lower;
+        ts = (static_cast<int64_t>(ts_upper) << 32) |
+              static_cast<uint32_t>(ts_lower);
     }
 
     // Check the flags to see if there is another 24 bytes of timestamp
